Moved Exchanges File into BitCoin part, create getWallets method
