--- src/main/scala/tcs/custom/ethereum/Exchange.scala
@@ -1,8 +1,0 @@
-package tcs.custom.ethereum
-
-class Exchange(name: String){
-
-  override def toString: String = {
-    "Exchange: " + name
-  }
-}
