+++ src/main/scala/tcs/examples/ethereum/Tries.scala
@@ -2,15 +2,11 @@ package tcs.examples.ethereum
 
 import org.web3j.protocol.Web3j
 import tcs.blockchain.ethereum.EthereumBlockchain
-import tcs.custom.ethereum.Exchanges
+import tcs.custom.bitcoin.Exchanges
 
 
 object Tries {
   def main(args: Array[String]): Unit = {
-    val infuraBlockchain = new EthereumBlockchain("http://localhost:8545/")
-      .setStart(1).setStep(100)
-    infuraBlockchain.foreach((block) => {
-      println(block.hash + ";" + block.number + ";" + block.transactions.length)
-    })
+    println(Exchanges.getWallets("Bittrex.com", 10012))
   }
 }
